Chore: move field name to module
