+++ public/app/features/logs/components/LogRowMessageDisplayedFields.test.tsx
@@ -4,8 +4,8 @@ import userEvent from '@testing-library/user-event';
 import { createTheme, LogLevel } from '@grafana/data';
 import { IconButton } from '@grafana/ui';
 
-import { LOG_LINE_BODY_FIELD_NAME } from './LogDetailsBody';
 import { LogRowMessageDisplayedFields, Props } from './LogRowMessageDisplayedFields';
+import { LOG_LINE_BODY_FIELD_NAME } from './fieldSelector/logFields';
 import { getLogRowStyles } from './getLogRowStyles';
 import { createLogRow } from './mocks/logRow';
