+++ public/app/features/logs/components/fieldSelector/getSuggestedFieldsFromTable.ts
@@ -1,10 +1,10 @@
 import { DataFrame } from '@grafana/data';
 import { getDistinctLabels } from 'app/features/transformers/utils';
 
-import { LOG_LINE_BODY_FIELD_NAME } from '../LogDetailsBody';
 import { getSuggestedOTelDisplayFormat, OTEL_LOG_LINE_ATTRIBUTES_FIELD_NAME } from '../otel/formats';
 
 import { FieldWithStats } from './FieldSelector';
+import { LOG_LINE_BODY_FIELD_NAME } from './logFields';
 
 export function getSuggestedFieldsFromTable(_: DataFrame, displayedFields: string[], defaultFields: string[] = []) {
   const suggestedFields: FieldWithStats[] = defaultFields.map((field) => ({
