+++ public/app/features/logs/components/fieldSelector/logFields.ts
@@ -1,0 +1,1 @@
+export const LOG_LINE_BODY_FIELD_NAME = '___LOG_LINE_BODY___';
