+++ public/app/features/logs/components/panel/LogLine.test.tsx
@@ -6,7 +6,7 @@ import { CoreApp, createTheme, getDefaultTimeRange, LogsDedupStrategy, LogsSortO
 import { config } from '@grafana/runtime';
 import { createTempoDatasource } from '@grafana-plugins/tempo/test/mocks';
 
-import { LOG_LINE_BODY_FIELD_NAME } from '../LogDetailsBody';
+import { LOG_LINE_BODY_FIELD_NAME } from '../fieldSelector/logFields';
 import { createLogLine } from '../mocks/logRow';
 import { getDisplayedFieldsForLogs, OTEL_PROBE_FIELD } from '../otel/formats';
