+++ public/app/features/logs/components/panel/LogLine.tsx
@@ -18,9 +18,9 @@ import { findHighlightChunksInText, GrafanaTheme2, LogsDedupStrategy, TimeRange 
 import { t } from '@grafana/i18n';
 import { Button, Icon, Tooltip } from '@grafana/ui';
 
-import { LOG_LINE_BODY_FIELD_NAME } from '../LogDetailsBody';
 import { LogLabels } from '../LogLabels';
 import { LogMessageAnsi } from '../LogMessageAnsi';
+import { LOG_LINE_BODY_FIELD_NAME } from '../fieldSelector/logFields';
 import { OTEL_LOG_LINE_ATTRIBUTES_FIELD_NAME } from '../otel/formats';
 
 import { HighlightedLogRenderer } from './HighlightedLogRenderer';
